+++ src/message_bot.py
@@ -5,9 +5,9 @@ import re
 import praw
 import prawcore
 
-from model import Database, MessageText, User
+from model import Database, MessageText, User, Flair
 from service import init_logger
-from text import parse_timezone, parse_day, parse_game, timezone_to_gmt, is_nsfw, sort_days
+from text import parse_timezone, parse_day, parse_game, timezone_to_gmt, is_nsfw, sort_days, find_all_keyword, parse_flair
 
 
 __reddit: praw.Reddit = praw.Reddit("message")
@@ -18,10 +18,12 @@ __production: bool = __reddit.config.custom["environment"] == "production"
 def read_messages(db: Database):
     for message in __reddit.inbox.stream():
         reply = parse_incoming_message(db, message)
-        if __production:
+        if __production ^ message.subject.endswith("devtesting"):
             message.mark_read()
             if reply:
                 message.reply(reply)
+        else:
+            __logger.info(f"Message to {message.author.name}\n{reply}")
         time.sleep(2)
 
 
@@ -44,14 +46,19 @@ def parse_incoming_message(db: Database, message: praw.models.Message) -> str:
     elif re.search(r'(bug|issue|error|feature|suggestion)', full_message, re.IGNORECASE):
         return MessageText.ERROR_REPLY
 
-    elif re.search(r'(sub(?:scribe)?|notify|lfg(?!\spost))', message.subject, re.IGNORECASE) or parse_game(message.subject):
+    elif re.search(r'(sub(?:scribe)?|notify|lfg(?!\spost))', message.subject, re.IGNORECASE) or parse_game(message.subject) or message.subject.endswith("devtesting"):
         game = parse_game(full_message)
         if not game:
             return MessageText.MISSING_GAME_REPLY
+        extra = ""
 
         user.game = game
         user.nsfw = is_nsfw(message.body)
         user.day = parse_day(message.body)
+        user.flair = parse_flair(message.body) or Flair.DEFAULT.flag
+        if user.flair and user.flair != Flair.DEFAULT.flag:
+            extra += f"- Flair (beta): {Flair.flag_to_str(user.flair)}  \n"
+
         timezone = parse_timezone(message.body)
         output = []
         if timezone:
@@ -59,14 +66,20 @@ def parse_incoming_message(db: Database, message: praw.models.Message) -> str:
             output = [f"{tz} ({timezone_to_gmt(tz)})" for tz in timezone]
             user.timezone = corrected
 
+        keywords = find_all_keyword(message.body)
+        if keywords:
+            user.keyword = '|'.join([re.escape(keyword) for keyword in keywords])
+            extra += f"""- Keyword{'s' if len(keywords) > 1 else ''} (beta): "{'" or "'.join(keywords)}"  \n"""
+
         user.save(db)
 
         return ("You have been successfully subscribed to LFG Notify Bot.  \n"
                 "&nbsp;  \n"
                 "Your current settings are:  \n"
-                f"- Game(s): {', '.join(user.game)}  \n"
-                f"- Timezone(s): {', '.join(output) if output else 'None Input'}  \n"
-                f"- Day(s) of the week: {', '.join(sort_days(user.day)) if user.day else 'None Input'}  \n"
+                f"- Game{'s' if len(user.game) > 1 else ''}: {', '.join(user.game)}  \n"
+                f"- Timezone{'s' if output and len(output) > 1 else ''}: {', '.join(output) if output else 'None Input'}  \n"
+                f"- Day{'s' if user.day and len(user.day) > 1 else ''} of the week: {', '.join(sort_days(user.day)) if user.day else 'None Input'}  \n"
+                f"{extra}"
                 f"- Include NSFW: {'Yes' if user.nsfw else 'No'}  \n"
                 "&nbsp;  \n"
                 "If you wish to change these settings, reply to this message (include all settings, not just your updates), or reply **STOP** to end notifications.  \n"