+++ src/text/__init__.py
@@ -1,5 +1,5 @@
 from .day_parser import parse_day, sort_days
-from .flag_parser import players_wanted, is_online, is_offline, is_lgbt, is_nsfw, is_one_shot, age_limit, using_vtt
+from .flag_parser import is_online, is_offline, is_lgbt, is_nsfw, is_one_shot, age_limit, using_vtt, find_all_keyword, is_play_by_post, parse_flair
 from .game_parser import parse_game
 from .time_parser import parse_time
 from .timezone_parser import parse_timezone, timezone_to_gmt
@@ -9,5 +9,5 @@ __all__ = [
     "parse_day", "sort_days",
     "parse_game",
     "parse_time",
-    "players_wanted", "is_online", "is_offline", "is_lgbt", "is_nsfw", "is_one_shot", "age_limit", "using_vtt"
+    "is_online", "is_offline", "is_lgbt", "is_nsfw", "is_one_shot", "age_limit", "using_vtt", "find_all_keyword", "is_play_by_post", "parse_flair"
 ]